AVEL_align/cdvae-align-feature-sly-align/requirements.txt
dtw
fastdtw
h5py
ipdb
librosa
pysptk
pysptk
pyworld
pyyaml
scipy
